48 - Refactor and handle network failure
